requirements.txt
Flask>=3.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
gunicorn
<<<<<<< HEAD

=======
>>>>>>> 1c874ed (Primer commit proyecto Bitácora Sofía)
